Arrumando service

+++ src/service/ClassePersonagemService.java
@@ -1,10 +1,12 @@
 package service;
 
 import entities.ClassePersonagem;
+import entities.Jogador;
 import entities.Personagem;
 import exceptions.BancoDeDadosException;
 import repository.ClassePersonagemRepository;
 
+import java.util.Objects;
 
 
 public class ClassePersonagemService {
@@ -12,7 +14,13 @@ public class ClassePersonagemService {
     ClassePersonagemRepository classePersonagemRepository = new ClassePersonagemRepository();
 
     public void adicionarClassePersonagem(Personagem personagem, ClassePersonagem classePersonagem) throws BancoDeDadosException {
-        classePersonagemRepository.adicionar(classePersonagem, personagem.getId());
+        if(personagem == null){
+            System.out.println("Personagem inexistente");
+        }else if(classePersonagem == null){
+            System.out.println("Classe inexistente");
+        }else if (verificaNomeClasse(classePersonagem)) {
+            classePersonagemRepository.adicionar(classePersonagem, personagem.getId());
+        }
     }
 
     public void listarTodos() throws BancoDeDadosException {
@@ -22,32 +30,40 @@ public class ClassePersonagemService {
     }
 
     public void remover(ClassePersonagem classePersonagem) throws BancoDeDadosException {
-        classePersonagemRepository.remover(classePersonagem.getIdClassePersonagem());
+        if(classePersonagem == null){
+            System.out.println("Classe inexistente");
+        }else {
+            classePersonagemRepository.remover(classePersonagem.getIdClassePersonagem());
+        }
     }
 
     public void editar(ClassePersonagem classePersonagem) throws BancoDeDadosException {
-        classePersonagemRepository.editar(classePersonagem.getIdClassePersonagem(),classePersonagem);
+        if(classePersonagem == null){
+            System.out.println("Classe inexistente");
+        }else if (verificaNomeClasse(classePersonagem)) {
+            classePersonagemRepository.editar(classePersonagem.getIdClassePersonagem(), classePersonagem);
+        }
     }
 
-//    public ClassePersonagem retornaClasse(Personagem personagem, String nome) throws BancoDeDadosException {
-//        return classePersonagemRepository.listar().stream().filter(c -> c.getIdPersonagem() == personagem.getId()).map( cp -> {
-//            return new ClassePersonagem()
-//                }
-//        );
-//    }
+    public ClassePersonagem retornaClasse(String nome) throws BancoDeDadosException {
+        return classePersonagemRepository.listar().stream().filter(a -> Objects.equals(a.getNomeClassePersonagem(), nome)).map(a -> {
+            return new ClassePersonagem(a.getIdClassePersonagem(), a.getNomeClassePersonagem(), a.getVidaClasse(), a.getDefesaClasse(), a.getAtaqueClasse());
+        }).findFirst().orElse(null);
+    }
 
+
+    public boolean verificaNomeClasse(ClassePersonagem classePersonagem) throws BancoDeDadosException {
+        if (classePersonagem == null) {
+            System.out.println("Classe inexistente");
+        } else {
+            ClassePersonagem classePersonagem1 = this.retornaClasse(classePersonagem.getNomeClassePersonagem());
+            if (classePersonagem1 == null) {
+                return true;
+            } else {
+                return false;
+            }
+        }
+        return false;
+    }
 }
-
-
-//    public Personagem retornaPersonagem(String nome) throws BancoDeDadosException {
-//        return personagemRepository.listar().stream().filter(a -> Objects.equals(a.getNomePersonagem(), nome)).map(a -> {
-//            return new Personagem(a.getId(),a.getNomePersonagem());
-//        }).findFirst().get();
-//
-//    }
-
-//    public void editar(Personagem personagem,String nome) throws BancoDeDadosException {
-//        personagem.setNomePersonagem(nome);
-//        personagemRepository.editar(personagem.getId(),personagem);
-//    }
