+++ src/service/PersonagemService.java
@@ -12,39 +12,57 @@ public class PersonagemService {
     PersonagemRepository personagemRepository = new PersonagemRepository();
 
     public void adicionar(Jogador jogador, Personagem personagem) throws BancoDeDadosException {
-        if(this.verificaNomePersonagem(personagem)) {
+        if (personagem == null) {
+            System.out.println("Personagem inexistente");
+        } else if (jogador == null) {
+            System.out.println("Jogador inexistente");
+        } else if (this.verificaNomePersonagem(personagem)) {
             personagemRepository.adicionar(personagem, jogador.getId());
         }
     }
 
-    public void editar(Personagem personagem,String nome) throws BancoDeDadosException {
-        personagem.setNomePersonagem(nome);
-        if(this.verificaNomePersonagem(personagem)){
-            personagemRepository.editar(personagem.getId(),personagem);
+    public void editar(Personagem personagem, String nome) throws BancoDeDadosException {
+        if (personagem == null) {
+            System.out.println("Personagem inexistente");
+        } else {
+            personagem.setNomePersonagem(nome);
+            if (this.verificaNomePersonagem(personagem)) {
+                personagemRepository.editar(personagem.getId(), personagem);
+            }
         }
     }
 
     public void listar() throws BancoDeDadosException {
-        for (Personagem personagem : personagemRepository.listar()){
+        for (Personagem personagem : personagemRepository.listar()) {
             System.out.println(personagem);
         }
     }
 
     public void remover(Personagem personagem) throws BancoDeDadosException {
-        personagemRepository.remover(personagem.getId());
+        if (personagem == null) {
+            System.out.println("Personagem inexistente");
+        } else {
+            personagemRepository.remover(personagem.getId());
+        }
     }
 
     public Personagem retornaPersonagem(String nome) throws BancoDeDadosException {
         return personagemRepository.listar().stream().filter(a -> Objects.equals(a.getNomePersonagem(), nome)).map(a -> {
-            return new Personagem(a.getId(),a.getNomePersonagem());
+            return new Personagem(a.getId(), a.getNomePersonagem());
         }).findFirst().orElse(null);
     }
+
     public boolean verificaNomePersonagem(Personagem personagem) throws BancoDeDadosException {
-        Personagem personagem1 = this.retornaPersonagem(personagem.getNomePersonagem());
-        if (personagem1 == null) {
-            return true;
+        if (personagem == null) {
+            System.out.println("Personagem inexistente");
         } else {
-            return false;
+            Personagem personagem1 = this.retornaPersonagem(personagem.getNomePersonagem());
+            if (personagem1 == null) {
+                return true;
+            } else {
+                return false;
+            }
         }
+        return false;
     }
 }
